Landing page created
-Webpack config reverted. Issue was not with pathing, but naming
convention casing. Currently using the default "index" instead of named
jsx files and letting the transpiler auto-detect. It was looking for
lower case and I had upper, which was causing a caching issue with
bundle.js

-Landing page to be remade later with a lighter background similar to
Figma Landing page concepts ExpressVPN style. Planning to use Dall-e to
generate a multi-layered background in this style.

-Consider changing font-style to a sans style.