Drag and Drop list in progress (WARNING)
WARNING: DO NOT PULL TO APP
The drag and drop list is being tested on the main #app element and so
the app itself is non-existant in current state. Forgot to branch.
Lesson learned.

Current state of DnD: Got the react component structure working for a
single linear ordered list. The ol component is now broken off and took
several hours to solve what turns out to be an 11 year old Chrome bug
requiring a setTimeout in the ondragstart event, for DOM manipulation.
It must have something to do with context, as the bug only appeared
after seperating out the ol into its own component.